+++ automagix/__init__.py
@@ -6,10 +6,10 @@ import sys
 from time import time, strftime, gmtime
 
 from .batch_runner import get_script_and_batch_items, run_batch_items
-from .config import init_logger, CONFIG, LOG, VERSION, arguments, MAGIC_SELECTION_INT
+from .config import init_logger, CONFIG, LOG, VERSION, arguments, MAGIC_SELECTION_INT, progress_bar
 from .helpers import empty_queued_input_data, selector
 from .parallel_runner import run_parallel_screens
-from .progress_bar import setup_scroll_area, destroy_scroll_area
+
 
 def run_startup_script():
     if not CONFIG.get('startup_script'):
@@ -82,13 +82,10 @@ def main():
         sys.exit(0)
 
     try:
-        if CONFIG['progress_bar']:
-            setup_scroll_area()
-
+        progress_bar.setup()
         run_batch_items(script=script, batch_items=batch_items, args=args)
     finally:
-        if CONFIG['progress_bar']:
-            destroy_scroll_area()
+        progress_bar.destroy()
 
     if 'AUTOMAGIX_TIME' in os.environ:
         LOG.info(f'The Automagix script took {round(time() - starttime)}s!')