+++ automagix/command.py
@@ -7,8 +7,8 @@ from shlex import quote
 from time import time
 
 from .colors import italic, yellow, green, red
+from .config import progress_bar
 from .environment import PipelineEnvironment, AttributedDict, AttributedDummyDict, ConversionError
-from .progress_bar import draw_progress_bar
 
 PERSISTENT_VARS = PVARS = AttributedDict()
 
@@ -188,8 +188,7 @@ class Command:
             # _ask_user handles are answers but PA.retry, PA.skip, PA.proceed
             # PA.retry and PA.proceed are not in allowed options
             # PA.skip means 'skip' so we can just go on
-        if self.env.config['progress_bar']:
-            draw_progress_bar(self.progress_portion)
+        progress_bar.draw(self.progress_portion)
 
     def _execute(self, interactive: bool = False, force: bool = False):
         self.print_command()
@@ -304,7 +303,7 @@ class Command:
         formatted_options = options.format(bash_path=self.bash_path)
 
         return self._ask_user_with_options(
-            question=f'{question}\n{formatted_options}\nYour answer: \a',
+            question=f'{question}\n{formatted_options}\nYour answer: ',
             allowed_options=allowed_options,
         )