+++ automagix/environment.py
@@ -3,9 +3,8 @@ from logging import getLogger
 from pathlib import Path
 from typing import Any
 
-from .config import init_logger
+from .config import init_logger, progress_bar
 from .helpers import empty_queued_input_data
-from .progress_bar import block_progress_bar, draw_progress_bar
 
 
 class AttributedDict(dict):
@@ -78,15 +77,13 @@ class PipelineEnvironment:
         return
 
     def interact(self, question: str, progress_portion: int = None) -> str:
-        if progress_portion is not None and self.config['progress_bar']:
-            block_progress_bar(progress_portion)
+        progress_bar.block(percentage=progress_portion, cursor_col=len(question.split('\n')[-1]))
         self.send_status('user_input_add')
         empty_queued_input_data()
-        print(question, end='')
+        print(question, end='\a')
         answer = input()
         self.send_status('user_input_remove')
-        if progress_portion is not None and self.config['progress_bar']:
-            draw_progress_bar(progress_portion)
+        progress_bar.draw(percentage=progress_portion, cursor_col=0)
         return answer
 
     def set_var(self, key: str, value: Any) -> None: